+++ src/components/organisms/reviewsWithInfiniteScroll/index.tsx
@@ -7,7 +7,7 @@ import { getReviews } from "@/effect/reviews/getReviews";
 import { FilterBar } from "@/components/molecules/filterBar";
 import { Filters } from "@/entity/filters";
 import { reviewsSortOptions } from "@/constants/sortOptions";
-import { getFormattedDate } from "@/libs/date/getFormattedDate";
+import { getFormattedDate } from "@/shared/lib";
 
 interface ReviewDetailProps {
   initialReviews: ReviewDetail[];
